refactor: factor utilities into utils.js
