+++ lib/moonbeam.js
@@ -10,6 +10,12 @@ const uuid = require('uuid/v4')
 const dfns = require('date-fns')
 const isValidAccount = require('eos-name-verify')
 
+const {
+  sendGenericError,
+  formatCandles,
+  formatTrades
+} = require('./utils')
+
 const Sunbeam = require('sunbeam')
 
 const { Api, JsonRpc } = require('eosjs')
@@ -31,10 +37,6 @@ const schema = {
 
 const log = (...args) => {
   console.error.apply(null, args)
-}
-
-const dec = (n) => {
-  return +n.toFixed(8)
 }
 
 class Moonbeam {
@@ -378,10 +380,6 @@ class Moonbeam {
   onPrivateHistoryHttpRequest (req, res) {
     const { maxResults } = this.conf
 
-    const sendGenericError = () => {
-      res.status(500).json({ error: 'ERR_GENERIC' })
-    }
-
     const payload = req.body
     const _v = Joi.validate(payload, schema)
     if (_v.error) {
@@ -411,7 +409,7 @@ class Moonbeam {
           return res.status(500).json({ error: 'ERR_WS_DOWN' })
         }
 
-        return sendGenericError()
+        return sendGenericError(res)
       }
 
       const [ meta, valid ] = vres
@@ -436,7 +434,7 @@ class Moonbeam {
         .sort({ ts: -1 })
         .project({ 'entry': 1, _id: 0 })
         .toArray((err, entries) => {
-          if (err) return sendGenericError()
+          if (err) return sendGenericError(res)
 
           const cleaned = entries.map((el) => {
             return el.entry
@@ -446,34 +444,6 @@ class Moonbeam {
         })
     })
   }
-}
-
-function formatCandles (candles) {
-  const res = candles.map((entry) => {
-    return [
-      entry.t,
-      dec(entry.open),
-      dec(entry.close),
-      dec(entry.high),
-      dec(entry.low),
-      entry.volume
-    ]
-  })
-
-  return res
-}
-
-function formatTrades (trades) {
-  const res = trades.map((entry) => {
-    return [
-      entry.id,
-      entry.t,
-      entry.amount,
-      entry.price
-    ]
-  })
-
-  return res
 }
 
 function server (opts, plugins) {
